pyproject.toml
[tool.poetry]
name = "E-COM test task"
version = "0.1.0"
description = "Test task for e-com backend developer"
authors = ["Bilibin Sergey"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
motor = "^3.6.0"
fastapi = "^0.115.6"
pydantic-settings = "^2.6.1"
uvicorn = "^0.32.1"
pydantic-extra-types = "^2.10.1"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
autoflake = "^2.3.1"
flake8-pyproject = "^1.2.3"
mypy = "^1.13.0"
types-python-dateutil = "^2.9.0.20241206"

[tool.poetry.group.test.dependencies]
httpx = "^0.28.1"

[tool.flake8]
max-line-length = 120
extend-ignore = ["E203", "C901"]

[tool.black]
line-length = 120

[tool.isort]
line_length = 120
multi_line_output = 3

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"